swicth to bfloat16 instead of float16 for GPU compatibility

+++ app/components/llm.py
@@ -27,7 +27,7 @@ def load_llm():
                 load_in_4bit=True,
                 bnb_4bit_use_double_quant=True,
                 bnb_4bit_quant_type="nf4",
-                bnb_4bit_compute_dtype=torch.float16,
+                bnb_4bit_compute_dtype=torch.bfloat16,
             )
             model = AutoModelForCausalLM.from_pretrained(
                 HUGGINGFACE_REPO_ID,
