+++ train/evaluate.py
@@ -58,7 +58,7 @@ def load_model_pipeline(adapter_path: str | None = None):
         load_in_4bit=True,
         bnb_4bit_use_double_quant=True,
         bnb_4bit_quant_type="nf4",
-        bnb_4bit_compute_dtype=torch.float16,
+        bnb_4bit_compute_dtype=torch.bfloat16,
     )
     tokenizer = AutoTokenizer.from_pretrained(MODEL_ID, token=HF_TOKEN)
     tokenizer.pad_token = tokenizer.eos_token
